Bulk_Article_dump/requirements.txt.txt
fastapi==0.75.0
uvicorn==0.17.0
mysql-connector-python==8.0.27
tqdm==4.62.3